+++ test/unit/fuzz-tools.test.js
@@ -97,6 +97,8 @@ import * as hemoncV94 from '../../lib/hemonc-v94.js';
 import * as neuroV95 from '../../lib/neuro-v95.js';
 import * as psychV96 from '../../lib/psych-v96.js';
 import * as periopV97 from '../../lib/periop-v97.js';
+import * as pedsV98 from '../../lib/peds-v98.js';
+import * as idcritV99 from '../../lib/idcrit-v99.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -138,6 +140,8 @@ const MODULES = {
   'neuro-v95.js': neuroV95,
   'psych-v96.js': psychV96,
   'periop-v97.js': periopV97,
+  'peds-v98.js': pedsV98,
+  'idcrit-v99.js': idcritV99,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];